Basic Network Object

The basic Pathway Commons network. This is all SIF Pathway commons interactions loaded into a network. The CHEMI interactions are present. The igraph network object consisting of 30,910 genes and 1,902,605 interactions. basic_network$graph is a list of networks. Each network is every interaction in Pathway Commons consisting of a specific type of interaction.

basic_network

Format

A list object

igraph::V(basic_network$network)

Vertex names

igraph::E(basic_network$network)

Network edges

igraph::E(basic_network$network)$interaction

Type of interaction

igraph::E(basic_network$network)$Occurance

How many Pathway Commons databases record an interaction between to and from vertex

igraph::E(basic_network$network)$UniqCol

The Unique Code of the entire interaction

igraph::E(basic_network$network)$pathway

The Pathway(s) in Pathway commons this interaction is documented in as a list object

igraph::E(basic_network$network)$EdgeRep

How many interaction types are there between to and from vertacies

igraph::E(basic_network$network)$Edge

The to:from name desigantion of the edge

igraph::E(basic_network$network)$SumOccurancel

how many times is to vertex connected to from vertex across all pathway commons data bases and interaction types

...