Traces the shortest paths of a gene to a vector of Target Gene and a vector of Sentinal Genes

Traces the the shortest paths of target gene paiwise to the target gene list. then traces the shortest path of the gene to a list of sentinal genes. This function returns the list of genes in a path object which score over the an emperically defined limit using the find_limit() function.

short_paths(tnet, target, targets, sentinals, cores = 1)

Arguments

tnet
igraph network (Main entire network) eg. net/net_undirected/JS_net_undirected

target
the from gene target eg Genes[1]

targets
List of the total list of targets in the User set eg. Genes

sentinals
List of the sentinal genes to trace to eg. Sentinal

cores
the number of cores to use in paralellizable functions

Value

List object of Inter genes from target path traces and Sentinal genes from sentinal gene traces

Examples

 
data(slim_net, package = "igraphNetworkExpansion")
data(genelists, package = "igraphNetworkExpansion")

example_path <- short_paths(
    tnet = slim_net,
    target = genelists$targets$APP_Metabolism[1],
    targets = genelists$targets$APP_Metabolism,
    sentinals = genelists$sentinals$Immune_Response,
    cores = 1
)

#> Working on: GRIN2B

#> 313 target paths out of 536 ( 58.4% ) kept

#> 42 of 137 target genes ( 30.66% ) filtered out

#> 316 sentinel paths out of 646 ( 48.92% ) kept

#> 48 of 180 sentinel genes ( 26.67% ) filtered out