Test and example network: implicit loading

A Full Template network to use for testing and examples. All protein-protein interactions in pathway commons for genes exressed in at least one tissue from AMP-AD transcriptomics data.

slim_net

Format

An igraph network object

verticies

genes

edges

protien interactions

igraph::V(slim_net$name)

The Vertex Gene Name

igraph::V(slim_net$community)

A Sample igraph Community designation for the vertex

igraph::V(slim_net$weight)

A Sample Weight Value

igraph::V(slim_net$RNA_EffectScore)

A Sample RNA Phenotype Effect Score

igraph::V(slim_net$Pro_EffectScore)

A Sample Protien Phenotype Effect Score

igraph::V(slim_net$RNA_Cor_EffectScore)

A Sample RNA co-expression value

igraph::V(slim_net$Pro_Cor_EffectScore)

A Sample Protien co-expression value

igraph::E(slim_net$interaction)

The Edge Interaction Type

igraph::E(slim_net$UniqCol)

The From_vert:To_vert-Interaction_type unique identifier

igraph::E(slim_net$Edge)

The From_vertex:To_vertex identifier

igraph::E(slim_net$pathway)

The Pathway Commons data base(s) where this edge is found

igraph::E(slim_net$EdgeRep)

How many interaction types are there between the two vertices across pathway commons databases

igraph::E(slim_net$Occurance)

How many Data Bases across pathway commons record an interaction of anytype between the two vertices

igraph::E(slim_net$SumOccurence)

Total amount of times these to vertices interact accross all pathway commons database and interaction types

...