Push Network to Synapse

This function takes a network object and pushes it to synapse. The provided example will work assuming that you have edit privliges to the synapse ID used as the patent ID (p_id).

store_net(
  network,
  net_filename,
  net_synname,
  p_id,
  folder,
  act_name,
  act_desc,
  synap_import,
  client_import,
  code = NULL,
  repo = NULL,
  syn_used = NULL,
  subset = NULL,
  prov_object = NULL
)

Arguments

network
the igraph network to push to synapse eg. net

net_filename
the file name of the network without file extension

net_synname
the desplay name of the network in synapse

p_id
the parent synapse ID of the network destination

folder
the name of the storage folder in the parent synapse ID to store the net

act_name
the name of the syn activity object to

act_desc
the description of the syn activity object to

synap_import
is the reticulated imported synapse from log_into_synapse()$synapse

client_import
is the reticulated imported synapseclient client log_into_synapse()$client

code
the path of the code which generated the network for the provenance (optional)

repo
the repo which generated the network for the provenance (optional)

syn_used
character vector of synIDs to seed the provenance (optional)

subset
An vector of vertex names to filter the network for (optional) eg. test

prov_object
A pre made github code provenance object or vector of objects

Value

a synapse entity of a .graphml subnetwork object stored in synapse

Examples

if (FALSE) {
data(slim_net, package = "igraphNetworkExpansion")
syn <- log_into_synapse()

store_net(
    network  = slim_net , 
    net_filename = "test_netpush", 
    net_synname = "test_netpush_network",
    p_id = "syn25467038", 
    folder = "test_push",
    act_name = "This is a test push network",
    act_desc = "This network was puched here to test the store_net function",
    synap_import = syn$synapse, 
    client_import = syn$client
) 
}